cmd/link/internal/ld/pe: Fix segfault adding resource section
When `rsrc.P` is used directly, the array backing the byte slice seems
to be at a very odd area in memory which causes the code to later
segfault when assigning to the slice (see associated issue for segfault
address). This change fixes it so that the linker no longer segfaults
when accessing the slice.

Fixes #42384